Move all classes into internal package
