+++ bundles/org.openhab.binding.upb/src/main/java/org/openhab/binding/upb/internal/UPBController.java
@@ -20,10 +20,9 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.upb.UPBDevice;
-import org.openhab.binding.upb.UPBDevice.DeviceState;
-import org.openhab.binding.upb.handler.UPBThingHandler;
-import org.openhab.binding.upb.handler.VirtualThingHandler;
+import org.openhab.binding.upb.internal.UPBDevice.DeviceState;
+import org.openhab.binding.upb.internal.handler.UPBThingHandler;
+import org.openhab.binding.upb.internal.handler.VirtualThingHandler;
 import org.openhab.binding.upb.internal.message.Command;
 import org.openhab.binding.upb.internal.message.UPBMessage;
 import org.openhab.binding.upb.internal.message.UPBMessage.Type;
