+++ bundles/org.openhab.binding.upb/src/main/java/org/openhab/binding/upb/internal/UPBDevice.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.upb;
+package org.openhab.binding.upb.internal;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
