+++ bundles/org.openhab.binding.upb/src/main/java/org/openhab/binding/upb/internal/UPBHandlerFactory.java
@@ -24,10 +24,9 @@ import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.eclipse.smarthome.io.transport.serial.SerialPortManager;
-import org.openhab.binding.upb.Constants;
-import org.openhab.binding.upb.handler.SerialPIMHandler;
-import org.openhab.binding.upb.handler.UPBThingHandler;
-import org.openhab.binding.upb.handler.VirtualThingHandler;
+import org.openhab.binding.upb.internal.handler.SerialPIMHandler;
+import org.openhab.binding.upb.internal.handler.UPBThingHandler;
+import org.openhab.binding.upb.internal.handler.VirtualThingHandler;
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
