+++ bundles/org.openhab.binding.upb/src/main/java/org/openhab/binding/upb/internal/handler/UPBThingHandler.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.upb.handler;
+package org.openhab.binding.upb.internal.handler;
 
 import static org.openhab.binding.upb.internal.message.Command.*;
 
@@ -32,9 +32,9 @@ import org.eclipse.smarthome.core.thing.type.ChannelTypeUID;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
-import org.openhab.binding.upb.Constants;
-import org.openhab.binding.upb.UPBDevice;
-import org.openhab.binding.upb.handler.UPBIoHandler.CmdStatus;
+import org.openhab.binding.upb.internal.Constants;
+import org.openhab.binding.upb.internal.UPBDevice;
+import org.openhab.binding.upb.internal.handler.UPBIoHandler.CmdStatus;
 import org.openhab.binding.upb.internal.message.MessageBuilder;
 import org.openhab.binding.upb.internal.message.UPBMessage;
 import org.slf4j.Logger;
